+++ csv/applications/csv-time-join.cpp
@@ -251,7 +251,7 @@ int main( int ac, char** av )
         if( select_only && timestamp_only ) { std::cerr << "csv-time-join: --timestamp-only specified with --select, ignoring --timestamp-only" << std::endl; }
         bool discard_bounding = options.exists( "--discard-bounding" );
         boost::optional< unsigned int > buffer_size = options.optional< unsigned int >( "--buffer" );
-        if( options.exists( "--bound" ) ) { bound = boost::posix_time::microseconds( options.value< double >( "--bound" ) * 1000000 ); }
+        if( options.exists( "--bound" ) ) { bound = boost::posix_time::microseconds( static_cast<unsigned int>(options.value< double >( "--bound" ) * 1000000 )); }
         stdin_csv = comma::csv::options( options, "t" );
 
         std::vector< std::string > unnamed = options.unnamed(
