+++ io/applications/io-cat.cpp
@@ -323,7 +323,7 @@ int main( int argc, char** argv )
         std::string connect_max_attempts_string = options.value< std::string >( "--connect-max-attempts,--connect-attempts,--attempts,--max-attempts", "1" );
         connect_max_attempts = connect_max_attempts_string == "unlimited" ? 0 : boost::lexical_cast< unsigned int >( connect_max_attempts_string );
         double connect_period_seconds = options.value( "--connect-period", 1.0 );
-        connect_period = boost::posix_time::milliseconds( std::floor( connect_period_seconds * 1000 ) );
+        connect_period = boost::posix_time::milliseconds( static_cast<unsigned int>(std::floor( connect_period_seconds * 1000 ) ));
         permissive = options.exists( "--permissive" );
         const std::vector< std::string >& unnamed = options.unnamed( "--permissive,--exit-on-first-closed,-e,--flush,--unbuffered,-u,--verbose,-v", "-.+" );
         #ifdef WIN32
